Added EP Tables along an Axis
Added the ability to make tables (in csv format) based on the electric potential along the cardinal axis.

Also Added corrections to the way electric potential is calculated. Now electric potential is scaled universally based on the electric potential at the center of the wire grid (users can input the center or it defaults to the origin)