Fixed Inheritance Bug
Particle class extends Charge class, and originally both classes had a position vector and a polarity number. This redundancy was causing errors when particles being treated like charges by methods were unable to find their own variable. The fix was to remove the redundancy. Another bug has shown up, however, as when using the "plusEqual" method with numbers very close to zero for hte vector (think 1E-27), when adding a scalar component of a vector that is very small in magnitude to a scalar that is relatively large in magnitude, the number defaults to the smaller one (or so it appears) further research is needed.