Added Energy to Particles
Particles now have the following three properties (all doubles): Electric Potential Energy, Kinetic Energy, and Total Energy. Kinetic Energy can be calculated by the particle itself. However, Electric Potential Energy, and by extension, Total Energy, can only be calculated with the aid of the electric field. The calculations are methods in both the eField class and in the particle class, but to perform the calculation in the eField class you must pass it a charge/particle, and to perform the calculations in the particle class you must pass the eField