Unit Conversion Fixes
There were a number of unit related bug fixes. The two most prominent were a lack of taking into account the charge of a particle with one extra electron or proton, which was assumed to be 1 Coulomb instead of it's accurate value.

Secondly, particles that utilized EField's functions failed to take into account the scaleDistance. This was rectified. In future, units may be converted to solely meters, so as to avoid any other unit conversion confusion